Moved Login and Signup forms and fixed exports/imports
